backend/portfolio-upload/requirements.txt
boto3==1.35.82
psycopg2-binary==2.9.10
requests==2.32.3